🛠️ 🏗️ refactor: extract search bloom and create loading logic into dedicated modules
